add a delay to long discord messages
